bug fix upload picture and fix int to decimal

+++ BnhsQrCode/Model/ScanHistory.cs
@@ -9,7 +9,7 @@ namespace BnhsQrCode.Model
     {
         public virtual int Id { get; set; }
         public virtual int UserProfileId { get; set; }
-        public virtual int BodyTemp { get; set; }
+        public virtual decimal BodyTemp { get; set; }
         public virtual string TimeStatus { get; set; }
         public virtual DateTime CreatedByDateTime { get; set; }
     }
